+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 ### Fixed
 * Focus now automatically moves to the hour input after a date is selected from the picker.